backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic==2.9.0
pydantic-settings>=2.0.0
apscheduler==3.10.4
httpx==0.27.0
beautifulsoup4==4.12.3
html2text==2024.2.26
sentence-transformers==3.3.0
rapidfuzz==3.10.0
python-dotenv==1.0.1
aiofiles==24.1.0
openai==1.57.0
python-jobspy==1.1.82
sqlite-vec==0.1.6
python-multipart==0.0.17
pandas==2.2.0
playwright==1.49.0
python-docx==1.1.0
fpdf2==2.8.0
pylatexenc==2.10
pdfplumber==0.11.0
